bpf: migrate tail call maps to avoid altering currently loaded ones
We currently only migrate maps upon property mismatch when we reload BPF
programs, but we also need to do this for tail call maps. The currently
loaded ones should not be altered as this results in weird behavior until
the main prog has been replaced. Thus even if there was no property
mismatch, move the pinned tail call map to a different location and then
upon success unlink it. Remove the 'unlinking migrated node' message
since it is not worthy to warn here.

Fixes: #13015
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>